Update README.md
Taking https://github.com/lo1tuma/eslint-plugin-mocha/issues/120 it seems to make sense to add this small reference in the "Install and configure" section. Regardless of mocha (and not the plugin) being responsible for this, the plugin is in a position to provide more-than-enough information and save troubles to implementing developers.